requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cleanup==5.2.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.2.0
psycopg2-binary==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
